stash cleanArchi added
