+++ ozkour-back/infrastructure/googlesheets/sheets.js
@@ -1,7 +1,7 @@
 const { google } = require('googleapis')
-const connect = require('./connect.js')
+const connect = require('../googleslide/connect.js')
 const dayjs = require('dayjs')
-const utilitary = require('../Utils/dateUtils')
+const utilitary = require('../../Utils/dateUtils')
 const customParseFormat = require('dayjs/plugin/customParseFormat')
 dayjs.extend(customParseFormat)
 
@@ -10,28 +10,9 @@ dayjs.extend(customParseFormat)
  * @param {String} start the start of the date range
  * @param {String} end the end of the date range
  */
-async function getTalkFromDate (start, end = dayjs()) {
-  const formatedDateStart = dayjs(start)
-  const formatedDateEnd = dayjs(end)
-  if (formatedDateStart.format('MM/YYYY') === formatedDateEnd.format('MM/YYYY')) {
-    const param = { start: formatedDateStart.format('DD/MM/YYYY'), end: formatedDateEnd.format('DD/MM/YYYY') }
-    const res = await connect.authMethode(getData, param)
-    return res
-  } else {
-    let res = []
-    let tempDateStart = formatedDateStart
-    let tempDateEnd = formatedDateStart.add(1, 'month')
-    while (tempDateStart.format('MM/YYYY') !== formatedDateEnd.format('MM/YYYY')) {
-      const param = { start: tempDateStart.format('DD/MM/YYYY'), end: tempDateEnd.format('DD/MM/YYYY') }
-      res = res.concat(await connect.authMethode(getData, param))
-      tempDateStart = tempDateStart.add(1, 'month')
-      tempDateEnd = tempDateStart.add(1, 'month')
-    }
-    const param = { start: tempDateStart.format('DD/MM/YYYY'), end: formatedDateEnd.format('DD/MM/YYYY') }
-    res = res.concat(await connect.authMethode(getData, param))
-
-    return res
-  }
+async function getTalkFromDate (param) {
+  const res = await connect.authMethode(getData, param)
+  return res
 }
 
 /**
