+++ ozkour-back/infrastructure/googleslide/connect.js
