+++ ozkour-back/tests/sheets.test.js
@@ -1,4 +1,4 @@
-const sheet = require('../google-api/sheets')
+const sheet = require('../infrastructure/googlesheets/sheets')
 test('convert date in month written', () => {
   // sheet.dateFilter(talks, '19/01/2021', '21/01/2021')
   const result = sheet.dateFilter(talks, '19-01-2021', '25-01-2021')
